add authorization to order enpoint, clean up order line item additions

+++ OrdersDomain.Core/Aggregates/Entities/Orders/LineItem.cs
@@ -10,8 +10,7 @@ namespace OrdersDomain.Core.Aggregates.Entities.Orders
         public string ItemName { get; set; }
         public string ItemDescription { get; set; }
         public int Quantity { get; set; }
-        public int BaseAmount { get; set; }
-
+        public decimal BaseAmount { get; set; }
         public decimal TaxAmount { get; set; }
         public decimal TotalAmount { get; set; }
     }
